hide enrollment menu
