+++ rw-legacy/lib/get-menu-option.js
@@ -3,7 +3,7 @@ function for returning menu options from a given option menu
 */
 
 var getDistrictBundles = require('../dat/district-bundles');
-
+var msgs = require('./msg-retrieve'); 
 module.exports = function(menu_option, menu_table,districtId){
     var districtBundles = getDistrictBundles();
     console.log('district bundles:  ' + JSON.stringify({bundles: districtBundles}))
@@ -35,6 +35,8 @@ module.exports = function(menu_option, menu_table,districtId){
         
     }else {
 
+        var isDistrictClosed = require('./isDistrictClosed');
+        var clientDistrict = JSON.parse(state.vars.client_json)? JSON.parse(state.vars.client_json).DistrictId : null;
         var table = project.getOrCreateDataTable(menu_table);
         var cursor = table.queryRows({'vars': {'option_number': menu_option}});
         if((!cursor.hasNext())){
@@ -45,6 +47,10 @@ module.exports = function(menu_option, menu_table,districtId){
             admin_alert = require('./admin-alert');
             admin_alert('Error in retrieving menu option - duplicate options\nTable name : ' + menu_table + '\nOption number : ' + menu_option);
             //throw 'ERROR: Duplicate options - Table name : ' + menu_table + ' - Option number : ' + menu_option;
+        }
+        if(clientDistrict && (option.vars.option_name == 'enr_order_start' || option.vars.option_name == 'enr_order_review_start') && isDistrictClosed(clientDistrict)) {
+            global.sayText(msgs('enrollment_ended', {}, service.vars.lang));
+            return null;
         }
         outstr = option.vars.option_name || option.vars.bundle_name || option.vars.input_name;
         return outstr;
